Refresh and LoadMore

+++ app/src/main/java/com/example/zhong/lasagna/home/HomeJB.java
@@ -23,7 +23,7 @@ public class HomeJB {
      */
 
     private boolean hasvisible;
-    private int previous_cursor;
+    private long previous_cursor;
     private long next_cursor;
     private int total_number;
     private int interval;
@@ -43,11 +43,11 @@ public class HomeJB {
         this.hasvisible = hasvisible;
     }
 
-    public int getPrevious_cursor() {
+    public long getPrevious_cursor() {
         return previous_cursor;
     }
 
-    public void setPrevious_cursor(int previous_cursor) {
+    public void setPrevious_cursor(long previous_cursor) {
         this.previous_cursor = previous_cursor;
     }
